+++ grunt/jshint.js
@@ -1,0 +1,7 @@
+module.exports = {
+    all: [
+        'dist/ember-data-django-rest-adapter.js',
+        'tests/adapter_tests.js',
+        'tests/adapter_embedded_tests.js'
+    ]
+}
