+++ grunt/karma.js
@@ -1,0 +1,5 @@
+module.exports = {
+    unit: {
+        configFile: 'karma.conf.js'
+    }
+}
