+++ grunt/neuter.js
@@ -1,0 +1,7 @@
+module.exports = {
+    build: {
+        basePath: 'src/',
+        src: 'src/main.js',
+        dest: 'dist/ember-data-django-rest-adapter.js'
+    }
+}
