+++ grunt/uglify.js
@@ -1,0 +1,6 @@
+module.exports = {
+    dist: {
+        src: 'dist/ember-data-django-rest-adapter.prod.js',
+        dest: 'dist/ember-data-django-rest-adapter.min.js'
+    }
+}
